fix(imports): update catalog import paths after refactoring
- Fix src/importers/catalog.ts to import from ../catalog instead of ../packages/catalog
- Fix src/importers/catalog.test.ts to use correct import path
- Resolves CI build failure caused by missing module references

All tests passing (90/90), build successful (76.6kb)
